CommFactorCalc/src/test/java/za/co/sammysa/AppTest.java
package za.co.sammysa;

import org.junit.Test;
import static org.junit.Assert.*;

import junit.framework.Test;
import junit.framework.TestCase;
import junit.framework.TestSuite;

/**
 * Unit test for simple App.
 */
public class AppTest 
    extends TestCase
{
    /**
     * Create the test case
     *
     * @param testName name of the test case
     */
    public AppTest( String testName )
    {
        super( testName );
    }

    /**
     * @return the suite of tests being tested
     */
    public static Test suite()
    {
        return new TestSuite( AppTest.class );
    }

    /**
     * Rigourous Test :-)
     */
    public void testApp()
    {
        assertTrue( true );
    }

    @Test
    public void testFindHighestCommonFactor()
    {
        int[] numbers = {1, 2, 4, 8, 16};
        assertEquals(16, App.findHighestCommonFactor(numbers));
    }
}
